Fix UpdateHutContainsItems check for item categories
Commit ed0dbe134754 ("Fixed most issues for 1.6", 2024-03-24) updated
the BetterJunimos logic for Stardew Valley 1.6

When it modified the check for item categories in
UpdateHutContainsItems(), it broke the handling for item categories.

In Stardew Valley 1.6, item IDs all got converted to strings, and the
commit tried to fix this by simply changing the check from 'itemId < 0'
to 'itemId != "0"'

This effectively broke item category handling. As a result, in
IdentifyJunimoAbility, the shortcut to check for item categories never
passes. This results in Junimos never performing many of their actions
such as planting crops.

Notice that all item categories are still prefixed with a '-'. The
correct check for splitting between categories and items is to check
that the ID in question has a '-' at the start. Replace the 'itemId !=
"0"' with a simple string StartsWith check. This fixes the logic and
allows storing the categories into the dictionary appropriately.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>