extract ItemsInHuts update logic to UpdateHutContainsItemId
The UpdateHutContainsItems() function has a helper for updating item
categories in UpdateHutContainsItemCategory. This feels a bit awkward
since we have the logic for updating items directly in this function.

We could pull the logic for category update into this function, but that
makes it less legible due to the itemId variable name

Instead, lets extract the regular itemId code to its own helper
function, UpdateHutContainsItemId.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>