Update logic for Junimo Hut cost reduction to Stardew 1.6
In Stardew Valley 1.6, the Data/Blueprints object was removed and
replaced with Data/Buildings, which represents the building data using a
BuildingData object class.

Switch the asset load request to edit Data/Buildings instead of
Data/Blueprints, and edit it as a list of BuildingData.

We no longer need to split or join with '/' and can just edit the
BuildingData fields directly. Update the material list to match the
reduced cost as the previous blue print logic.

We can no longer just display the blueprint data as a string, so update
the monitor debug logs accordingly.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>